fix(radio display): radio display
